+++ index.html
@@ -85,7 +85,7 @@
         <script src="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.4.1/semantic.min.js"></script>
         <script src="https://unpkg.com/flickity@2/dist/flickity.pkgd.min.js"></script>
         <script src="./public/js/flickity.js"></script>
-        <script src="./public/js/scipt.js"></script>
+        <script src="./public/js/script.js"></script>
       </body>
       </html>
